requirements.txt
Django==4.2.10
django-filter==24.2
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
drf-spectacular==0.27.2
